moved attention to models; deleted two CPPs
