--- src/layers/generic.cpp
@@ -1,1 +1,0 @@
-// TODO: delete this
