--- src/layers/guided_alignment.cpp
@@ -1,1 +1,0 @@
-// TODO: delete this
