+++ src/models/attention.cu
@@ -1,11 +1,11 @@
-#include "rnn/attention.h"
+#include "attention.h"
 
 #include "graph/node_operators_binary.h"
 #include "kernels/tensor_operators.h"
 
 namespace marian {
 
-namespace rnn {
+namespace models {
 
 struct AttentionNodeOp : public NaryNodeOp {
   AttentionNodeOp(const std::vector<Expr>& nodes)
