+++ src/models/s2s.h
@@ -3,6 +3,7 @@
 #include "marian.h"
 #include "layers/constructors.h"
 #include "rnn/constructors.h"
+#include "constructors.h"
 
 namespace marian {
 
@@ -214,7 +215,7 @@ private:
           auto encState = state->getEncoderStates()[k];
 
           baseCell.push_back(
-              rnn::attention(graph)("prefix", attPrefix).set_state(encState));
+              models::attention(graph)("prefix", attPrefix).set_state(encState));
         }
       }
     }
@@ -312,7 +313,7 @@ public:
       auto att = rnn_->at(0)
                      ->as<rnn::StackedCell>()
                      ->at(k + 1)
-                     ->as<rnn::Attention>();
+                     ->as<models::Attention>();
       alignedContexts.push_back(att->getContext());
     }
 
@@ -364,7 +365,7 @@ public:
   // helper function for guided alignment
   virtual const std::vector<Expr> getAlignments(int i = 0) {
     auto att
-        = rnn_->at(0)->as<rnn::StackedCell>()->at(i + 1)->as<rnn::Attention>();
+        = rnn_->at(0)->as<rnn::StackedCell>()->at(i + 1)->as<models::Attention>();
     return att->getAlignments();
   }
