+++ src/rnn/constructors.h
@@ -3,7 +3,6 @@
 #include "marian.h"
 #include "layers/factory.h"
 #include "rnn/rnn.h"
-#include "rnn/attention.h"
 
 namespace marian {
 namespace rnn {
@@ -133,32 +132,6 @@ public:
 };
 
 typedef Accumulator<StackedCellFactory> stacked_cell;
-
-// TODO: move attention and this factory to models/
-class AttentionFactory : public InputFactory {
-protected:
-  Ptr<EncoderState> state_;
-
-public:
-  AttentionFactory(Ptr<ExpressionGraph> graph) : InputFactory(graph) {}
-
-  Ptr<CellInput> construct() {
-    ABORT_IF(!state_, "EncoderState not set");
-    return New<Attention>(graph_, options_, state_);
-  }
-
-  Accumulator<AttentionFactory> set_state(Ptr<EncoderState> state) {
-    state_ = state;
-    return Accumulator<AttentionFactory>(*this);
-  }
-
-  int dimAttended() {
-    ABORT_IF(!state_, "EncoderState not set");
-    return state_->getAttended()->shape()[1];
-  }
-};
-
-typedef Accumulator<AttentionFactory> attention;
 
 class RNNFactory : public Factory {
 protected:
